requirements.txt
faster-whisper>=1.0.3
pytesseract>=0.3.10
Pillow>=10.2.0
opencv-python>=4.9.0.80
numpy>=1.26.0
ImageHash>=4.3.1
google-cloud-vision>=3.7.2
assemblyai>=0.37.0
